Move recyclerview ext
